requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
arize-ai-grhub-dataset-embeddings-guardrails==0.0.0
arize-otel==0.9.0
arrow==1.3.0
attrs==25.3.0
banks==2.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
diff-match-patch==20230430
dirtyjson==1.0.8
distro==1.9.0
Faker==25.9.2
filelock==3.18.0
filetype==1.2.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.7.0
fuzzysearch==0.8.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
griffe==0.36.9
grpcio==1.73.1
guardrails-ai==0.6.6
guardrails-api-client==0.4.0a1
guardrails-grhub-ban-list==0.1.0
guardrails-grhub-valid-length==0.0.0
guardrails-grhub-web-sanitization==0.0.0
guardrails_hub_types==0.0.4
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
isoduration==20.11.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain-core==0.3.69
langsmith==0.4.7
lark==1.2.2
litellm==1.74.4
llama-cloud==0.1.32
llama-cloud-services==0.6.43
llama-index==0.12.49
llama-index-agent-openai==0.4.12
llama-index-cli==0.4.4
llama-index-core==0.12.49
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.10
llama-index-instrumentation==0.3.0
llama-index-llms-openai==0.4.7
llama-index-multi-modal-llms-openai==0.5.3
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.11
llama-index-readers-llama-parse==0.4.0
llama-index-workflows==1.1.0
llama-parse==0.6.43
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.3.1
openai==1.97.0
openinference-instrumentation==0.1.35
openinference-instrumentation-guardrails==0.1.9
openinference-semantic-conventions==0.1.21
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-instrumentation==0.56b0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.11.0
packaging==25.0
pandas==2.2.3
pillow==11.3.0
platformdirs==4.3.8
propcache==0.3.2
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
pydash==7.0.7
Pygments==2.19.2
PyJWT==2.10.1
pypdf==5.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rpds-py==0.26.0
rstr==3.2.2
safetensors==0.5.3
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
tryolabs-grhub-restricttotopic==0.1.0
typer==0.15.4
types-python-dateutil==2.9.0.20250708
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.0.7
webcolors==24.11.1
webencodings==0.5.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0